Moved PilotConfig class into pilot
